styling of form and move structure
