+++ src/components/compound_form/compound_form.scss
@@ -1,0 +1,14 @@
+
+
+.form__group {
+  width: 50%;
+}
+
+.form__label {
+  display: block;
+}
+
+.form__input {
+  display: block;
+  width: 100%;
+}
